Refactor pump and water management system.
- Updated pump pin assignments to use defined constants for clarity.
- Improved initialization of pump modes with clearer comments.
- Enhanced the Pump class constructor for better readability.
- Removed unused EEPROM read/write functions and replaced them with a struct-based approach for configuration management.
- Implemented a new Configuration struct to store all settings, including pump amounts and durations.
- Refactored water management to utilize AppState for threshold values instead of EEPROM.
- Added electrovalve control functionality to manage water flow more effectively.
- Updated functions to save and load configuration data, ensuring valid settings are applied.
- Cleaned up commented-out code and improved overall code organization for maintainability.

+++ appstate.cpp
@@ -9,10 +9,10 @@
 #include "appstate.h"
 
 namespace AppState {
-  uint8_t languageIndex = 0;
-  Pump pumps[PUMP_COUNT];
-  uint32_t tankVolume = 0;
-  int64_t timeOffset = 0;
-  uint8_t lowThreshold = 100;
-  uint8_t highThreshold = 0;
+uint8_t languageIndex = 0;
+Pump pumps[PUMP_COUNT];
+uint32_t tankVolume = 0;
+int64_t timeOffset = 0;
+uint16_t lowThreshold = 100;
+uint16_t highThreshold = 0;
 }
